Adds support for solarized.
Currently supports the 16 color version of the theme as recommended by
altercation.

The functionality required that I remove csapprox which broke the
terminal theme quite badly.

I also added the functionality to auto-set the background to dark or
light based on whether I was in the terminal or in a GUI.