Stop syntastic from auto-opening the errors list
This patch prevents syntastic from automatically opening the errors
window as soon as it finds errors. This should also allow us to close
popup windows for files that still have a lot of errors that have not
yet been cleaned up.

It also cleans up some comments on this section to make it clearer what
settings are doing what.